OpenPandora Port from Sebastien Chevalier: * created a new branch for this port * email from  Sebastien Chevalier:
Hi,

This mail to inform you that I have ported both Jedi Knight 2 and Jedi Knight 3 to the OpenPandora.
The Pandora is a Handheld computer running Linux on ARM, and using GLES for Hardware acceleration.

I converted all GL code to GLES, and added some ARM and NEON (the MMX/SSE coporocessor for ARM) specific code.

You can see the resulting binary here: http://repo.openpandora.org/?page=detail&app=jedioutcast_ptitseb and here: http://repo.openpandora.org/?page=detail&app=jediacademy_ptitseb

And two thread in the official Boad here: http://boards.openpandora.org/index.php/topic/13621-jedi-knight-ii-jedi-outcast/ and here: http://boards.openpandora.org/index.php/topic/13664-jedi-knight-3-jedi-academy/

[...]

Sebastien.

+++ code/cgame/FxScheduler.cpp
@@ -28,7 +28,7 @@ void CMediaHandles::operator=(const CMediaHandles &that )
 {
 	mMediaList.clear();
 
-	for ( int i = 0; i < that.mMediaList.size(); i++ )
+	for ( int i = 0; i < int(that.mMediaList.size()); i++ )
 	{
 		mMediaList.push_back( that.mMediaList[i] );
 	}
@@ -225,7 +225,7 @@ int CFxScheduler::RegisterEffect( const char *file, bool bHasCorrectPath /*= fal
 	}
 
 	// If we'll overflow our buffer, bail out--not a particularly elegant solution
-	if (len >= sizeof(data) - 1 )
+	if (len >= int(sizeof(data)) - 1 )
 	{
 		theFxHelper.CloseFile( fh );
 		return 0;
@@ -1409,7 +1409,7 @@ void CFxScheduler::CreateEffect( CPrimitiveTemplate *fx, vec3_t origin, vec3_t a
 		height = fx->mHeight.GetVal();
 
 		// calculate point on ellipse
-		VectorSet( temp, sin(x) * width * sin(y), cos(x) * width * sin(y), cos(y) * height ); // sinx * siny, cosx * siny, cosy
+		VectorSet( temp, sinf(x) * width * sinf(y), cosf(x) * width * sinf(y), cosf(y) * height ); // sinx * siny, cosx * siny, cosy
 		VectorAdd( org, temp, org );
 
 		if ( fx->mSpawnFlags & FX_AXIS_FROM_SPHERE )
