+++ code/cgame/cg_consolecmds.cpp
@@ -256,7 +256,7 @@ qboolean CG_ConsoleCommand( void ) {
 
 	cmd = CG_Argv(0);
 
-	for ( i = 0 ; i < sizeof( commands ) / sizeof( commands[0] ) ; i++ ) {
+	for ( i = 0 ; i < int(sizeof( commands ) / sizeof( commands[0] )) ; i++ ) {
 		if ( !Q_stricmp( cmd, commands[i].cmd ) ) {
 			commands[i].function();
 			return qtrue;
@@ -278,7 +278,7 @@ so it can perform tab completion
 void CG_InitConsoleCommands( void ) {
 	int		i;
 
-	for ( i = 0 ; i < sizeof( commands ) / sizeof( commands[0] ) ; i++ ) {
+	for ( i = 0 ; i < int(sizeof( commands ) / sizeof( commands[0] )) ; i++ ) {
 		cgi_AddCommand( commands[i].cmd );
 	}
 }
