+++ code/game/bg_pangles.cpp
@@ -39,7 +39,7 @@ void PM_ScaleUcmd( playerState_t *ps, usercmd_t *cmd, gentity_t *gent )
 	{//driving a vehicle
 		//clamp the turn rate
 		int maxPitchSpeed = MAX_PITCHSPEED_X_WING;//switch, eventually?  Or read from file?
-		int diff = AngleNormalize180(SHORT2ANGLE((cmd->angles[PITCH]+ps->delta_angles[PITCH]))) - floor(ps->viewangles[PITCH]);
+		int diff = AngleNormalize180(SHORT2ANGLE((cmd->angles[PITCH]+ps->delta_angles[PITCH]))) - floorf(ps->viewangles[PITCH]);
 	
 		if ( diff > maxPitchSpeed )
 		{
@@ -52,7 +52,7 @@ void PM_ScaleUcmd( playerState_t *ps, usercmd_t *cmd, gentity_t *gent )
 
 		//Um, WTF?  When I turn in a certain direction, I start going backwards?  Or strafing?
 		int maxYawSpeed = MAX_YAWSPEED_X_WING;//switch, eventually?  Or read from file?
-		diff = AngleNormalize180(SHORT2ANGLE(cmd->angles[YAW]+ps->delta_angles[YAW]) - floor(ps->viewangles[YAW]));
+		diff = AngleNormalize180(SHORT2ANGLE(cmd->angles[YAW]+ps->delta_angles[YAW]) - floorf(ps->viewangles[YAW]));
 
 		//clamp the turn rate
 		if ( diff > maxYawSpeed )
